update002

+++ app/Http/Controllers/AutorController.php
@@ -15,8 +15,7 @@ class AutorController extends Controller
     {
         //
         $autores = Autor::paginate(25);
-        // preciso responder o usuario
-        return view('admin.autor.index', compact('autores'));
+        return view('admin.autores.index', compact('autores'));
     }
 
     /**
@@ -25,11 +24,7 @@ class AutorController extends Controller
     public function create()
     {
         //
-        $autores = Autor::all();
-
-        return view('site.autor.create', compact('autores'));
-
-
+        return view('admin.autores.index');
     }
 
     /**
@@ -38,12 +33,9 @@ class AutorController extends Controller
     public function store(StoreAutorRequest $request)
     {
         //
-
-        //Aqui vamos tratar as regras de salvamento e  vamos persistir no banco 
         Autor::create($request->all());
-       //Redirecionar ou  devolver uma mensagem para o cliente
-       //return redirect()->route('/noticias.index');
-       return redirect()->away('/autores')->with('success','Autor  criado com sucesso!');
+        return redirect()->away('/autores')
+            ->with('success', 'Autor criado com sucesso!');
     }
 
     /**
@@ -52,16 +44,7 @@ class AutorController extends Controller
     public function show(Autor $autor)
     {
         //
-
-        //$id -> recebendo via api
-        // $noticia = Noticia::find($id);
-        // $nome e eu quero o primeiro registro
-        // $noticia = Noticia::where('nome',$nome)->first();
-        // $noticia = Noticia::where('nome',$nome)->get();
-        // $noticia = Noticia::where('nome',$nome)->paginate();
         return view('admin.autores.show', compact('autor'));
-
-
     }
 
     /**
@@ -70,9 +53,7 @@ class AutorController extends Controller
     public function edit(Autor $autor)
     {
         //
-
-        $autores = Autor::all();
-        return view('admin.autor.edit', compact('autores'));
+        return view('admin.autores.edit', compact('autor'));
     }
 
     /**
@@ -81,9 +62,9 @@ class AutorController extends Controller
     public function update(UpdateAutorRequest $request, Autor $autor)
     {
         //
-
         $autor->update($request->all());
-        return redirect()->away('/autores')->with('success', 'Autores atualizado com sucesso!');
+        return redirect()->away('/autores')
+            ->with('success', 'Autor atualizado com sucesso!');
     }
 
     /**
@@ -92,12 +73,13 @@ class AutorController extends Controller
     public function destroy(Autor $autor)
     {
         //
+        if ($autor->noticias()->count() > 0) {
 
+            return redirect()->away('/autores')
+                ->with('success', 'Autor possui dependentes!');
+        }
         $autor->delete();
-
-        
-        return redirect()->away('/autores')->with('success', 'Deletado  com sucesso!');
-
-
+        return redirect()->away('/autores')
+            ->with('success', 'Autor criado com sucesso!');
     }
 }
