+++ app/Http/Controllers/CidadeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreCidadeRequest;
 use App\Http\Requests\UpdateCidadeRequest;
 use App\Models\Cidade;
+use App\Models\Estado;
 
 class CidadeController extends Controller
 {
@@ -14,11 +15,9 @@ class CidadeController extends Controller
     public function index()
     {
         //
-        //Carregue os dados do banco
-        //select * from cadernos
-        $cidade = Caderno::all();
-        // fazer o response pro usuario
-        return view('admin.cadernos.index',compact('cidades'));
+        $cidades = Cidade::paginate(25);
+
+        return view('admin.cidades.index', compact('cidades'));
     }
 
     /**
@@ -26,9 +25,9 @@ class CidadeController extends Controller
      */
     public function create()
     {
-        $cidade = Cidade::all();
-        $estado = Cidade::all();
-        return view('site.cidades.create',compact('cidades','estados'));
+        //
+        $estados = Estado::all();
+        return view('admin.cidades.create', compact('estados'));
     }
 
     /**
@@ -37,12 +36,9 @@ class CidadeController extends Controller
     public function store(StoreCidadeRequest $request)
     {
         //
-        //Aqui vamos tratar as regras de salvamento e  vamos persistir no banco 
         Cidade::create($request->all());
-       //Redirecionar ou  devolver uma mensagem para o cliente
-       //return redirect()->route('/noticias.index');
-       return redirect()->away('/cidades')->with('success','Cidade criada com sucesso!');
-
+        return redirect()->away('/cidades')
+            ->with('success', 'Cidade possui dependentes!');
     }
 
     /**
@@ -51,10 +47,7 @@ class CidadeController extends Controller
     public function show(Cidade $cidade)
     {
         //
-        
-        return view('admin.cadernos.show', compact('caderno'));
-
-
+        return view('admin.cidades.show', compact('cidade'));
     }
 
     /**
@@ -63,9 +56,8 @@ class CidadeController extends Controller
     public function edit(Cidade $cidade)
     {
         //
-        $cidades = Autor::all();
-        return view('admin.cidade.edit', compact('cidades'));
-
+        $estados = Estado::all();
+        return view('admin.cidades.edit', compact('cidade','estados'));
     }
 
     /**
@@ -74,10 +66,10 @@ class CidadeController extends Controller
     public function update(UpdateCidadeRequest $request, Cidade $cidade)
     {
         //
-        //
         $cidade->update($request->all());
-        return redirect()->away('/cidades')->with('success', 'Cidades  atualizados com sucesso!');
 
+        return redirect()->away('/cidades')
+            ->with('success', 'Cidade possui dependentes!');
     }
 
     /**
@@ -85,11 +77,13 @@ class CidadeController extends Controller
      */
     public function destroy(Cidade $cidade)
     {
-        //
+        if ($cidade->enderecos()->count() > 0) {
+            return redirect()->away('/cidades')
+                ->with('error', 'Cidade possui dependentes!');
+        }
+
         $cidade->delete();
-
-        
-        return redirect()->away('/cidades')->with('success', 'Deletado  com sucesso!');
+        return redirect()->away('/cidades')
+            ->with('success', 'Cidade removido com sucesso!');
     }
-
 }
